bin: let's ignore everything by default
Since this directory is used as the user binaries directory, some executables
might be installed from local builds. Since we don't want to include them in the
repository let's ignore everything there. Whenever a new script is to be added
on this repository I'll use the -f flag from `git add`.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>