emacs: properly check the OS and use that
The font detection was based on the window system, which utterly failed when
starting up on server mode. Thus, we should use the `system-type' variable
instead.

Moreover, I'm taking this oportunity to create two new constants which are
convenient to check the two supported OSes, and I'm also checking the OS at the
very beginning.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>