emacs: turns out that magit can indeed stage by patches
Before knowing this I had to resort to the CLI and manually `git stash -p` all
the way. Turns out that magit is actually able to stage particular changes from
a specific files, so this whole workflow is no longer needed, and hence the
warning on the documentation is no longer valid.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>