email: well, restrictions are stupid
The 'ajuntament' maildir has to be handled separately and manually
because resources there are scarce, and after 2GiB of space consumed, it
does not allow to proceed further.

Therefore, I am nuking everything on the server while preserving things
locally with a 'Pull New Flags' mode on mbsync. This has the side-effect
that now emails that have been already handled on the website and moved
into archive are no longer perceived to be moved, and the 'inbox' folder
might keep on growing. That is why I have set up a hook before mu4e's
update so these emails are removed every time.

Again, restrictions are stupid and I should not have to do this.

Signed-off-by: Miquel Sabaté Solà <msabate@suse.com>