emacs: drastically simplify my email setup
Instead of having a myriad of accounts, let's set up only two of them,
which are the actual ones I use on mu4e. This gets rid of hacks I did to
keep some of the accounts floating, while also dropping some shenanigans
like signatures which were never really relevant.

With all of this off loaded from Emacs, the msmtp configuration also
gets simplified, and I took the chance to tweak further some of the
defaults.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>