bin: Force a busybox path when running QEMU
When running QEMU/RISC-V, force the user to provide the `BUSYBOX_SRC`
environment variable. This variable points to the path where busybox has
been built. This way the path to the initramfs is not hardcoded.

Moreover, this allows us to rebuilt the initramfs every time the
`run-riscv-qemu.sh` is called, which in turn brings some consistency to
the process. Plus, this opens the door for allowing files to be passed
into a new `/home` directory for the current user. This way, someone can
just statically compile a binary via cross compilation, and then run it
in isolation with QEMU/RISC-V.

Last but not least, also update the .bashrc file with a value for
`BUSYBOX_SRC` on my system.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>