Streamline configuration for run-riscv-qemu.sh
Personal stuff like the location of the kernel image should go into my
.bashrc file, and then require it on `run-riscv-qemu.sh`. Similarly,
some environment variables were dropped, and some others removed. This
brings some much needed consistency to the whole script.

Last but not least, the `run-riscv-qemu.sh` now also had an extra
configurable variable, which is used for the `-cpu` QEMU option. It
defaults to 'max' so we can test the maximum amount of supported
extensions as possible.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>