emacs: Make "Inbox" a favorite bookmark
Apparently default notifications expect a favorite to be picked among
bookmarks. If not, then it picks the one at top, which is basically the
last one that was added by the user. Hence, define this bookmark last
just so it appears at the top of the main view, and also mark it as
"favorite" to double down.

Signed-off-by: Miquel Sabaté Solà <mssola@mssola.com>