kbuild: Change 'vf2' to 'debian'
The folks at openSUSE were nice enough to provide an image that works with my
VisionFive2 board. Hence, let's reformulate the "vf2" profile to "debian", as
that's closer to what I was doing.

That being said, I have left a TODO to denote that the UUID used in GRUB needs
to be handled more gracefully.

Signed-off-by: Miquel Sabaté Solà <mssola@mssola.com>